requirements.txt
Werkzeug==0.16.1
Flask==1.1.1
protobuf==3.15.6
gunicorn==20.0.4
pyOpenSSL==19.1.0
google-cloud-datastore==2.1.0
google-cloud-storage==1.36.2
